% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{EPIC_lncRNA_probes}
\alias{EPIC_lncRNA_probes}
\title{A character vector that maps CpG probes to lncRNA genes.
The vectors names are genes and the values are CpG probes}
\format{
A character vector with 205645 elements
}
\usage{
EPIC_lncRNA_probes
}
\description{
A character vector that maps CpG probes to lncRNA genes.
The vectors names are genes and the values are CpG probes
}
\examples{
\dontrun{
data("EPIC_lncRNA_probes")
}
}
\keyword{internal}
