% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{EPIC_miRNA_probes}
\alias{EPIC_miRNA_probes}
\title{A dataframe that maps CpG probes genes to microRNA genes.
The vectors names are genes and the values are CpG probes}
\format{
A dataframe with 23,379 rows and 3 columns
}
\usage{
EPIC_miRNA_probes
}
\description{
A dataframe that maps CpG probes genes to microRNA genes.
The vectors names are genes and the values are CpG probes
}
\examples{
\dontrun{
data("EPIC_miRNA_probes")
}
}
\keyword{internal}
