% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{EpigenomeMap}
\alias{EpigenomeMap}
\title{A list that map epigenome groups to epigenome IDs}
\format{
A list with 17 elements. The names are epigenome/tissue groups and the values are epigenome IDs for specific cells.
}
\usage{
EpigenomeMap
}
\description{
The data were generated from the RoadmapEpigenomic project (Nature, PMID: 25693563, figure 2).
}
\examples{
\dontrun{
data("EpigenomeMap")
}
}
\keyword{internal}
