% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{HM450_miRNA_probes}
\alias{HM450_miRNA_probes}
\title{A dataframe that maps CpG probes genes to microRNA genes.}
\format{
A dataframe with 17,495 rows and 4 columns
}
\usage{
HM450_miRNA_probes
}
\description{
A dataframe that maps CpG probes genes to microRNA genes.
}
\examples{
\dontrun{
data("HM450_miRNA_probes")
}
}
\keyword{internal}
