% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{lncRNA.data}
\alias{lncRNA.data}
\title{Toy lncRNA expression dataset for demonstration purpose.}
\format{
a matrix
}
\usage{
lncRNA.data
}
\description{
This is a subset of lncRNA expression data from TCGA-LUAD, used for testing the program.
}
\examples{
\dontrun{
data("lncRNA.data")
}
}
\keyword{internal}
