% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{LUAD.sample.annotation}
\alias{LUAD.sample.annotation}
\title{toy dataset for sample annotation}
\format{
a dataframe
}
\usage{
LUAD.sample.annotation
}
\description{
a dataframe with two columns: the first column is the patient identifier, and the second column is the sample annotation
}
\examples{
\dontrun{
data("LUAD.sample.annotation")
}
}
\keyword{internal}
