% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{MET.data}
\alias{MET.data}
\title{Toy DNA methylation dataset for demonstration purpose.}
\format{
a matrix
}
\usage{
MET.data
}
\description{
This is a subset of DNA methylation data from TCGA-LUAD, used for testing the program.
}
\examples{
\dontrun{
data("MET.data")
}
}
\keyword{internal}
