% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{microRNA.data}
\alias{microRNA.data}
\title{Toy microRNA expression dataset for demonstration purpose.}
\format{
a matrix
}
\usage{
microRNA.data
}
\description{
This is a subset of micoRNA expression data from TCGA-LUAD, used for testing the program.
}
\examples{
\dontrun{
data("microRNA.data")
}
}
\keyword{internal}
