% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{mRNA.data}
\alias{mRNA.data}
\title{Toy gene expression dataset for demonstration purpose.}
\format{
a matrix
}
\usage{
mRNA.data
}
\description{
This is a subset of gene expression data from TCGA-LUAD, used for testing the program.
}
\examples{
\dontrun{
data("mRNA.data")
}
}
\keyword{internal}
