% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{Sample_EpiMixResults_lncRNA}
\alias{Sample_EpiMixResults_lncRNA}
\title{toy dataset for EpiMix output with lncRNA mode}
\format{
a list
}
\usage{
Sample_EpiMixResults_lncRNA
}
\description{
a list object returned from EpiMix
}
\examples{
\dontrun{
data("Sample_EpiMixResults_lncRNA")
}
}
\keyword{internal}
