% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix.data.R
\docType{data}
\name{TCGA_BatchData}
\alias{TCGA_BatchData}
\title{A dataframe with the batch information of TCGA patient.}
\format{
a dataframe
}
\usage{
TCGA_BatchData
}
\description{
a dataframe with two columns: the first column is the patient identifier, and the second column is the technical batch
}
\examples{
\dontrun{
data("TCGA_BatchData")
}
}
\keyword{internal}
